code tidy

+++ app/src/main/java/com/example/second_course_starting_point/api/CharacterModel.kt
@@ -47,7 +47,7 @@ data class Wand(
     @SerializedName("core")
     val core: String,
     @SerializedName("length")
-    val length: Int // Double
+    val length: Double // Double
 
     //@SerializedName("boolean-key")
 //    val someBooleanReturnedFromJSON: Boolean
